requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
diskcache==5.6.3
filelock==3.16.1
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
llama_cpp_python==0.3.2
MarkupSafe==3.0.2
numpy==2.1.3
packaging==24.2
PyYAML==6.0.2
requests==2.32.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
